books now can be add, edit and delete

+++ app/Http/Controllers/BookController.php
@@ -7,25 +7,31 @@ use Illuminate\Http\Request;
 
 class BookController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except(['index', 'show']);
+    }
 
     public function index()
     {
-        // $books = Book::without('quizzes')->get();
-        $books = Book::all();
-        // return $books;
+        $books = Book::without('quizzes')->get();
         return view('books.index')->with(['books' => $books]);
     }
 
     public function create()
     {
-        // return view
+        $this->authorize('admin');
+        return view('Books.create');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('admin');
         $this->validate($request, [
             'title' => 'required',
+            'info' => 'required',
             'link' => 'required',
+            'img' => 'required',
         ]);
         $book = Book::create([
             'title' => $request->title,
@@ -33,37 +39,34 @@ class BookController extends Controller
             'link' => $request->link,
             'img' => $request->img,
         ]);
-        return $book;
-        // return show with book
+        return $this->show($book);
     }
 
     public function show(Book $book)
     {
         $book->load('quizzes');
-        // return $book;
+        //TODO: only 10 quiz??
         return view('books.show')->with(['book' => $book]);
     }
 
     public function edit(Book $book)
     {
-        // return form with book
+        $this->authorize('admin');
+        return view('books.edit')->with(['book' => $book]);
     }
 
     public function update(Request $request, Book $book)
     {
-        $book->update([
-            'title' => $request->title,
-            'info' => $request->info,
-            'link' => $request->link,
-            'img' => $request->img,
-        ]);
-        return $book;
-        //return show of book
+        $this->authorize('admin');
+        $input = collect(request()->all())->filter()->all();
+        $book->update($input);
+        return back();
     }
 
     public function destroy(Book $book)
     {
+        $this->authorize('admin');
         $book->delete();
-        // return index
+        return $this->index();
     }
 }
